Failing test: conflicting prices.
This is is a failing test to drive a refactor.

Currently, the price of the products is kept
on the products themselves, which are stored
inside the basket. This means that two baskets
with the same products could have different
prices if they were created at different times
with different product prices.

With this change, I am looking to push the
pricing of products to be centralized in the Store.

If a product isn't registered with the store
then they'll use the price listed on the product.